### 2019-10-04 #1784 @whedon
[PRE REVIEW]: pyHoops: A Python package for advanced basketball data analytics
**Submitting author:** @alessandroBombelli (<a href="http://orcid.org/0000-0001-7889-9552">Alessandro Bombelli</a>)
**Repository:** <a href="https://github.com/alessandroBombelli/pyHoops" target ="_blank">https://github.com/alessandroBombelli/pyHoops</a>
**Version:** v1.5
