### 2019-10-04 #1784 @kyleniemeyer
Hello @alessandroBombelli, thanks for your interest in JOSS. Before we proceed, during my review of your submission, I could not find the research application of your software. While useful for folks in basketball, I don't quite see where someone would cite this in the academic literature (not an explicit requirement, but one way we think about whether software has a research use).

Can you clarify this, per one of our submission requirements?
> The software should have an **obvious** research application.
