### 2019-10-04 #1784 @danielskatz
https://github.com/alessandroBombelli/pyHoops/pull/46 should fix the immediate paper building process
