### 2019-10-04 #1784 @alessandroBombelli
> Hello @alessandroBombelli, thanks for your interest in JOSS. Before we proceed, during my review of your submission, I could not find the research application of your software. While useful for folks in basketball, I don't quite see where someone would cite this in the academic literature (not an explicit requirement, but one way we think about whether software has a research use).
>
> Can you clarify this, per one of our submission requirements?
>
> > The software should have an **obvious** research application.

Dear @kyleniemeyer, apologies if the research application is not clear. I agree that, at a first glance, the obvious research application might not be as straight-forward as in other cases. On the other hand, a strong interest in data analytics applied to sports has been growing in recent years, as testified by ad-hoc conferences (MIT Sloan Sports Analytics Conference) and journals (Journal of Sports Sciences). To this avail, the research application of this work is the possibility to compute a set of team performance indices (per player and per lineup) to be used, as example, to infer lineup performances in future games. Knowing how different lineups perform against different opposing lineups (e.g., "tall", or "fast" lineups), can be used as part of a decision-making process to improve performances of a team. In addition, a sufficiently comprehensive database retrieved via pyHoops (e.g., a full season), can be used as the input for a prediction model (e.g., random forest) to assess what are the features (lineup- or player-specific) more crucial for a team's success. I hope this answers your question. I am willing to elaborate more if the answer is not satisfactory. Thank you!
