+++ paper.md
@@ -8,8 +8,11 @@ tags:
 authors:
   - name: Alessandro Bombelli
     orcid: 0000-0001-7889-9552
-    affiliation: Air Transport & Operations, Faculty of Aerospace Engineering, Delft University of Technology
-date: 05 October 2019
+    affiliation: 1 
+affiliations:
+ - name: Air Transport & Operations, Faculty of Aerospace Engineering, Delft University of Technology
+   index: 1
+date: 04 October 2019
 bibliography: pyHoops_references.bib
 ---
 
