### 2019-10-04 #1784 @alessandroBombelli
> [alessandroBombelli/pyHoops#46](https://github.com/alessandroBombelli/pyHoops/pull/46) should fix the immediate paper building process

Dear @danielskatz, sorry for the issue. but it is my first time here and, as such, I am not familiar with the formatting. I followed the example provided in the guidelines but thought that, being a single author, the distinction between affiliation and affiliations was not necessary. I fixed the issue now. Please let me know otherwise.
Thank you very much
