### 2019-10-04 #1784 @kyleniemeyer
@alessandroBombelli ok, so I think it would help to revise your article to better describe the research applications of the software, so this is clear to reviewers (and readers), focusing specifically on the types of research questions/problems this helps answer.

What you are describing to me sounds more like what a basketball team might use, but not an academic researcher writing a publication (for example)—but I may still be misunderstanding.
