### 2019-10-04 e58c36e
lots of fixes to language and to bib entries
