+++ paper.md
@@ -18,15 +18,15 @@ bibliography: pyHoops_references.bib
 
 # Summary
 
-Advanced data analytics has received contrasting opinions within the sport community. One of the greatest American Football of all times, Bill Belichick, stated that he uses analytics "less than zero" when preparing games [@belichick2019]. On the other end, the movie "Money Ball" stressed the necessity of advanced data analytics in baseball, i.e., sabermetric, to improve team's performances. Restricting the field of view to basketball, there exist, especially in the United States, a strong interest towards data science to enhance players' performances. Europe, on the other hand, is still a step behind in this sense. While statistics on individual performances of players are generally available after games in the form of box scores, such statistics are generally miopic. In fact, being individual statistics, they cannot fully grasp the impact of a player on his/her team while on the court and, equally importantly, the impact on the opposing team.
+Advanced data analytics has received contrasting opinions within the sport community. One of the greatest American Football coaches of all times, Bill Belichick, stated that he uses analytics "less than zero" when preparing games [@Belichick2019]. On the other end, the movie "Money Ball" stressed the necessity of advanced data analytics in baseball, i.e., sabermetrics, to improve team's performances. Restricting the field of view to basketball, there exist, especially in the United States, a strong interest in using data science to enhance players' performances. Europe, on the other hand, is still a step behind in this sense. While statistics on individual performances of players are generally available after games in the form of box scores, such statistics are generally miopic. In fact, being individual statistics, they cannot fully grasp the impact of a player on his/her team while on the court and, equally importantly, the impact on the opposing team.
 
-Basketball games in most European leagues and in Euroleague (the most important European basketball competition) are paired with a play-by-play report that maps every action (field goal made or attemped, foul, assist, substitution) of the game. Goal of ``pyHoops`` is to process such play-by-play reports to extrapolate performance indices that map the impact of players or lineups on the own and opposing team. As example, a player might be a prolific scorer, hence positively affecting his/her own offense, but a below-average defender, and thus negatively affect the team.
+Basketball games in most European leagues and in Euroleague (the most important European basketball competition) are paired with a play-by-play report that maps every action (field goal made or attemped, foul, assist, substitution) in the game. The goal of ``pyHoops`` is to process such play-by-play reports to extrapolate performance indices that map the impact of players or lineups on the both teams. As example, a player might be a prolific scorer, hence positively affecting his/her own offense, but a below-average defender, and thus negatively affect the team.
 
-``pyHoops`` is based on two major blocks: (1) web-parsing to translate play-by-play html tables into <code>pandas</code> databases, and (2) computation of statistics and performance indices. The web-parsing block easily stores play-by-play reports by automatically aceessing the associated webpage. Since webpages (and html tables) vary according to the specific league, the web-parsing block is league-specific and should be modified according to underlying html structure of the webpages of interest. As a last resort, play-by-play reports and boxscores can be directly copied from the webpage, saved as spreadsheets and loaded as databases to skip the web-parsing block. The secondo block uses play-by-play information to aggregate statistics for every distinct player and lineup, both for the own team and the opposing team.
[…]
