### 2019-10-04 #1784 @danielskatz
I've suggested more changes in https://github.com/alessandroBombelli/pyHoops/pull/48 - after merging this, you can enter ```@whedon generate pdf``` as a new comment here to regenerate the paper.
