### 2019-10-07 #1784 @danielskatz
@alessandroBombelli - do you have suggestions for reviewers?  They can be from [this list of potential reviewers](https://bit.ly/joss-reviewers) (please start at the bottom of the list), or individuals who are not on this list; if so, please mention their GitHub handles here.
