+++ paper.md
@@ -26,7 +26,7 @@ Basketball games in most European leagues and in Euroleague (the most important 
 
 Having now team-aggregated and not player-specific information, ``pyHoops`` can easily compute and compare, as example, the field goal percentage of the home team when a particular player was on the court, versus the field goal percentage of the away team when the same player was on the court. While each call of ``pyHoops`` focuses on a single game, season-wide statistics can be computed by running ``pyHoops`` everytime a new game occurs and storing new information in a <code>pandas</code> database, with the overarching goal of creating season-specific databases.
 
-We hope ``pyHoops`` can foster a better understanding and interest in advanced team-oriented basktetball statistics within the European basketball movement. Although some works already exist in the academic literature addressing advanced basketball analytics for European basketball [@travassos:2013;metulini:2018], we believe this work to be the first one (i) to focus on aggregated team-oriented statistics rather than individual-oriented statistics, and (ii) to provide an open-source package to be used be team data analysts of basketball enthusiasts to perform such analyses.      
+We hope ``pyHoops`` can foster a better understanding and interest in advanced team-oriented basktetball statistics within the European basketball movement. Although some works already exist in the academic literature addressing advanced basketball analytics for European basketball [@travassos:2013;@metulini:2018], we believe this work to be the first one (i) to focus on aggregated team-oriented statistics rather than individual-oriented statistics, and (ii) to provide an open-source package to be used be team data analysts of basketball enthusiasts to perform such analyses.      
 
 
 # Acknowledgements
