### 2019-10-08 #1784 @alessandroBombelli
> > I've suggested more changes in [alessandroBombelli/pyHoops#48](https://github.com/alessandroBombelli/pyHoops/pull/48) - after merging this, you can enter `@whedon generate pdf` as a new comment here to regenerate the paper.
>
> @danielskatz, I slightly modified references because I realized two of them were not correctly compiled. In one case, there was a @ missing. In the other case, where I am citing a website, I am not sure what the problem is. I am using the same citation type I would use in a classic LaTeX file, where I know it should work. On the other hand, I am not very familiar with the integration of the .md file with the .bib file, so I have little control on what is going wrong. Could you please help me with that? Thanks!

@danielskatz never mind, it is fixed now. Sorry for the confusion
