### 2019-10-08 #1784 @alessandroBombelli
> @alessandroBombelli - do you have suggestions for reviewers? They can be from [this list of potential reviewers](https://bit.ly/joss-reviewers) (please start at the bottom of the list), or individuals who are not on this list; if so, please mention their GitHub handles here.

Dear @danielskatz, I checked the list and I think @mado89 should be a good match. No preference for the second reviewer instead.
