### 2019-10-08 #1784 @danielskatz
Perhaps you can suggest some other people not on the list?  Maybe from other types of sport analytics?
