### 2019-10-08 #1784 @danielskatz
👋 @ctsilva - would you (or someone in your group) be willing to review this submission for JOSS? We carry out our checklist-driven reviews here in GitHub issues and follow these guidelines: https://joss.readthedocs.io/en/latest/review_criteria.html

If not, could you suggest some others in the field of software and sports analytics who might be suitable?
