### 2019-10-08 7b6bb73
Update pyHoops_references.bib
