### 2019-10-08 88a52f1
Update pyHoops_references.bib
