+++ pyHoops_references.bib
@@ -14,7 +14,7 @@
   number={1},
   pages={83--95},
   year={2013},
-  doi={\doi{10.1080/24748668.2013.11868633}}
+  doi={\doi{10.1080/24748668.2013.11868633}},
   publisher={Taylor \& Francis}
 }
 
