### 2019-10-08 65e9f89
Update pyHoops_references.bib
