### 2019-10-24 #1784 @mado89
What are the requirements for the reviewer? I might know a guy who knows somebody working in the industry  (sport data)Sent from my mobile devicePlease excuse briefness and typos

-------- Original message --------
