### 2019-10-27 #1784 @danielskatz
The reviewer should be familiar with open source, ideally with the domain and language of the software, and should not be conflicted with the authors
