### 2019-10-30 #1845 @whedon
[REVIEW]: pyHoops: A Python package for advanced basketball data analytics
**Submitting author:** @alessandroBombelli (<a href="http://orcid.org/0000-0001-7889-9552">Alessandro Bombelli</a>)
**Repository:** <a href="https://github.com/alessandroBombelli/pyHoops" target ="_blank">https://github.com/alessandroBombelli/pyHoops</a>
**Version:** v1.5

As of now there is no clear LICENSE file in the repository

No real installation instructions given. For me the software worked out of the box, but I assume it is only because I had coincidentally all required software installed.
No claims given so I cannot answer this. However, I noticed some problems and opened issues for them.

No real statement given
The
The provided test script provides some sample usage. However, the input and output could be explained better.
No thorough documentation of the script.
No tests. However, I think due to the nature of the software they might not be necessary. Hence, I ticked this one off.
The Readme contains the following statement "We welcome all feedback that might help improve the package.". I think this could be expanded to incorporate 1,2 and 3.

Not really. See comment below
Not really. See comment below
Not really. See comment below
see below
