### 2019-11-06 #1845 @danielskatz
👋 @mado89 & @ctsilva - I just wanted to check on how you are doing on your reviews, and to make sure that you are able to check boxes (and have accepted the invitation at https://github.com/openjournals/joss-reviews/invitations)
