### 2019-11-07 alessandroBombelli/pyHoops#49 @mado89
Missing LICENSE file
Although the readme states the project served under MIT license this file is missing (required by the journal conventions: "Does the repository contain a plain-text LICENSE file with the contents of an OSI approved software license?")
