### 2019-11-07 alessandroBombelli/pyHoops#49 @alessandroBombelli
> Although the readme states the project served under MIT license this file is missing (required by the journal conventions: "Does the repository contain a plain-text LICENSE file with the contents of an OSI approved software license?")
>
> [openjournals/joss-reviews#1845](https://github.com/openjournals/joss-reviews/issues/1845)

Hi @mado89, maybe I am missing something, but I added a MIT license I think ---> This project is licensed under the MIT License.
