### 2019-11-07 alessandroBombelli/pyHoops#49 @mado89
Yes that information is present. But afaik I understand it you need a copy of the license present in your repo and not only the reference.
Maybe @danielskatz can clarify.
