### 2019-11-07 #1845 @alessandroBombelli
> 👋 @mado89 & @ctsilva - I just wanted to check on how you are doing on your reviews, and to make sure that you are able to check boxes (and have accepted the invitation at https://github.com/openjournals/joss-reviews/invitations)

Dear @mado89 , @ctsilva, @danielskatz, I just realized the website I am using to web-scrape data for my examples changed the html structure of their play-by-play tables. This is retroactive, in the sense that is also applied to the webpages containing the scores I am using as examples. This means the examples will probably not work because I need to adjust the web-scraping issue. I sincerely apologize for this. I am planning to fix the problem over the weekend (best case scenario). I am not expecting the html structure to be changed anytime soon again, but I will find a different strategy for my examples so that if this situation occurs again, the code does not need to be fixed as it concerns the examples. Unfortunately, this is a problem that occurs with every web-scraping routine. If the original html webpage changes, the code must be adapted as well. If this is not consistent with your timeline or requirements, I totally understand. You can reject the paper and I will submit an improved version that circumvents this issue. My apologies
