### 2019-11-09 #1845 @mado89
Ok. Meanwhile, could you help me/us out over here? https://github.com/alessandroBombelli/pyHoops/issues/49
Question is whether its ok to state in the readme to be under MIT license or whether a LICENSE file is required
