### 2019-11-09 alessandroBombelli/pyHoops#49 @danielskatz
Yes, a LICENSE file is needed
