### 2019-11-12 #1845 @alessandroBombelli
> Sorry, I missed the notification from @alessandroBombelli 2 days ago.
>
> Let's just pause for now and see if this can be resolved fairly quickly.

Dear @mado89, @danielskatz, @ctsilva just a quick update. I should be done with my fixes in a couple of days tops. I learnt from this issue, and will generate a .csv of python pickle file that can be processed independently from the initial web-scraping routing, should the html structure of the webpage change anytime soon. Sorry again for the inconvenience
