+++ README.md
@@ -36,7 +36,7 @@ In the folder <copy>tests</copy>, the file <copy>pyHoops_tests.py</copy> can be 
 
 This project is licensed under the [MIT License](https://opensource.org/licenses/MIT).
 
-Copyright <2019> <Alessandro Bombelli>
+Copyright <2019,Alessandro Bombelli>
 
 Permission is hereby granted, free of charge, to any person obtaining a copy of this software and associated documentation files (the "Software"), to deal in the Software without restriction, including without limitation the rights to use, copy, modify, merge, publish, distribute, sublicense, and/or sell copies of the Software, and to permit persons to whom the Software is furnished to do so, subject to the following conditions:
 
