### 2019-11-13 #1845 @alessandroBombelli
> > Sorry, I missed the notification from @alessandroBombelli 2 days ago.
> > Let's just pause for now and see if this can be resolved fairly quickly.
>
> Dear @mado89, @danielskatz, @ctsilva just a quick update. I should be done with my fixes in a couple of days tops. I learnt from this issue, and will generate a .csv of python pickle file that can be processed independently from the initial web-scraping routing, should the html structure of the webpage change anytime soon. Sorry again for the inconvenience

Dear @mado89, @danielskatz, @ctsilva, I updated the code to reflect the changes in the original website I web-scrape. Tomorrow, I will update the github files so that they be downloaded and run. In the menawhile, I updated the readme.md file as well, and added the MIT license as requested. Thanks for the patience
