+++ pyHoops/__init__.py
@@ -1,24 +1,0 @@
-#!/usr/bin/env python3
-# -*- coding: utf-8 -*-
-"""
-Created on Thu Oct  3 17:36:35 2019
-
-@author: alessandro
-"""
-
-import numpy as np
-import requests
-from bs4 import BeautifulSoup
-import pandas as pd
-import os
-import matplotlib.pyplot as plt
-from matplotlib.cbook import get_sample_data
-from matplotlib.offsetbox import OffsetImage, AnnotationBbox
-import random
-import re
-import itertools
-import unicodedata
-from pyHoops import pyHoops
-
-            
-
