### 2019-11-14 b3b88de
Delete setup.py
