### 2019-11-14 0e1049f
Delete pyHoops.py
