+++ tests/pyHoops.py
@@ -1,2349 +1,0 @@
-#!/usr/bin/env python3
-# -*- coding: utf-8 -*-
-"""
-Created on Mon Sep 30 12:06:19 2019
-@author: alessandro
-"""
-
-#################################################
-### pyHoops: a Python package to web-parse
-### basketball games' play-by-play info
-### and, using basketball data analytics,
-### determine the effect of a player or a lineup
-### on the game
-#################################################
-
-import numpy as np
-import requests
-from bs4 import BeautifulSoup
-import pandas as pd
-import os
-import sys
-import matplotlib.pyplot as plt
-from matplotlib.cbook import get_sample_data
-from matplotlib.offsetbox import OffsetImage, AnnotationBbox
-import random
-import re
-import itertools
-import unicodedata
-
-# Input  : url pointing to the play-by-play webpage of the game of interest
-# Output : dataframe replicating the play-by-play table. To avoid issues with
-# special characters, accents, etc, they are replaced, while spaces are eliminated
-
-# Note: in the Italian Basketball Leaghe, the play-by-play table has three columns,
-# reporting respectively home team info, minute of the game or score, away team
-# info. For other leagues (e.g., Euroleague), ther emight be more columns,
-# and the web-parsing procedure should be slightly changed accordingly.
-
-
-
-def web_parse_playbyplay(thisUrl):
-    p    = requests.get(thisUrl)
-    
-    if p.status_code < 400:
-        l = []
-        soup = BeautifulSoup(p.text,'html.parser')
-        table = soup.find('table', attrs={'border':'0'})
-        
-        table_rows = table.find_all('tr')
-        for tr in table_rows:
-            td = tr.find_all('td')
-            row = [tr.text.lower() for tr in td]
-            if len(row)>0 and len(row)==3:
-                l.append(row)
-                
-    df = pd.DataFrame(np.array(l))
-    
-    
-    for i in range(0,int(len(df))):
-        for j in range(0,int(len(df.loc[i]))):
-            s             = df.loc[i][j]
-            s_mod         = ''.join((c for c in unicodedata.normalize('NFD', s) if unicodedata.category(c) != 'Mn'))
-            s_mod_noAp    = s_mod.replace('\'','')
-            s_mod_noSpace = s_mod_noAp.replace(' ','')
-            df.loc[i][j]  = s_mod_noSpace
-            
-            
-        
-    df.to_excel(l[0][0] + '_' + l[0][2] + '.xlsx',sheet_name='Sheet_name_1')
-    
-    return df,l[0][0],l[0][2] 
-
-# Input  : dataframe with play-by-play
-# Output : command window print of the game analyzed
-def print_game_analyzed(homeTeam,awayTeam):
-    # Python 3.X is running
-    if sys.version[0] == '3':
-        print('###########################################################')
-        print('Game analyzed is: ' + str(homeTeam) + ' vs ' + str(awayTeam))
-        print('###########################################################')
-
-# Input  : string identifying home team, string identifying away team, 
-#          list of available teams
-# Output : index of the logo that represents the home team and away team.
-#          If a team is not recognized in the list of teams available,
-#          it is assigned a default generic logo
-def determine_team_logo(string_homeTeam,string_awayTeam,team_list):
-    idx_logo_homeTeam = np.nan
-    for i in range(0,int(len(team_list))):
-        if team_list[i] in string_homeTeam:
-            idx_logo_homeTeam = i
-            break
-    if np.isnan(idx_logo_homeTeam):
-        idx_logo_homeTeam = int(len(team_list))
-    
-    idx_logo_awayTeam = np.nan
-    for i in range(0,int(len(team_list))):
-        if team_list[i] in string_awayTeam:
[…]
