### 2019-11-14 9db8d61
Delete pyHoops_tests.py
