+++ tests/pyHoops_tests.py
@@ -1,240 +1,0 @@
-#!/usr/bin/env python2
-# -*- coding: utf-8 -*-
-"""
-Created on Mon Sep 30 12:02:52 2019
-
-@author: alessandro
-"""
-
-import os
-import sys
-
-sys.path.append('C:/Python27/Lib/site-packages')
-sys.path.append('/Library/TeX/texbin/xelatex')
-sys.path.append('/anaconda/bin/python')
-sys.path.append('/anaconda/lib/python2.7/site-packages')
-sys.path.append('anaconda/envs/py37/lib/python3.7/site-packages')
-
-import pyHoops
-
-#######################
-### INPUTS:
-#######################
-
-# Path to current folder
-cwd             = os.getcwd()
-
-# List of teams and associated logos. An additional logo is added, to be associated
-# with teams that do not have a logo in the folder
-team_list = ['brescia','brindisi','cantu','cremona','fortitudobologna',
-'milano'	,'pesaro','pistoia','reggioemilia','roma','sassari','trentino',
-'treviso','trieste','varese','venezia','virtusbologna']
-
-team_list_logos = ['Brescia.png','Brindisi.png','Cantu.png','Cremona.png',
-                   'Fortitudo_Bologna.png','Milano.png'	,'Pesaro.png',
-                   'Pistoia.png','Reggio_Emilia.png','Roma.png','Sassari.png',
-                   'Trentino.png','Treviso.png','Trieste.png','Varese.png',
-                   'Venezia.png','Virtus_Bologna.png','LBA_logo.png'] 
-
-#%%
-#######################
-#######################
-### TESTS:
-#######################
-#######################
-
-# To run each test, uncomment the two urls:
-# 1) thisUrl        : url to play-by-play webpage of the game
-# 2) thisUrlBoxscore: url to boxscore webpage of the game
-# and run the code
-# IN the original version of this file, the game between
-# A|X Armani Exchange Milano and Germani Basket Brescia (TEST 3) is processed
-
-
-##################################################################
-### LBA Lega Basket Serie A: single game from 2018-2019 season ###
-##################################################################
-
-## TEST 1: Cantu - Avellino (2018-2019 season)
-#thisUrl         = 'http://web.legabasket.it/game/1672352/acqua_s_bernardo_cant__-sidigas_avellino_83:73/pbp'
-#thisUrlBoxscore = 'http://web.legabasket.it/game/1672352/acqua_s_bernardo_cant__-sidigas_avellino_83:73'
-
-##########################################################################
-### LBA Lega Basket Serie A: second week (all games), 2019-2020 season ###
-##########################################################################
-
-## TEST 2: Reggio Emilia - Trentino (2019-2020 season)
-#thisUrl         = 'http://web.legabasket.it/game/1672520/grissin_bon_reggio_emilia-dolomiti_energia_trentino_76:84/pbp'
-#thisUrlBoxscore = 'http://web.legabasket.it/game/1672520/grissin_bon_reggio_emilia-dolomiti_energia_trentino_76:84'
-
-# TEST 3: Armani Milano - Brescia (2019-2020 season)
-thisUrl         = 'http://web.legabasket.it/game/1672517/a_x_armani_exchange_milano-germani_basket_brescia-65:73/pbp'
-thisUrlBoxscore = 'http://web.legabasket.it/game/1672517/a_x_armani_exchange_milano-germani_basket_brescia-65:73'
-
-## TEST 4: Virtus Roma - Brindisi (2019-2020 season)
-#thisUrl         = 'http://web.legabasket.it/game/1672523/virtus_roma-happy_casa_brindisi_63:87/pbp'
[…]
