+++ tests/pyHoops.py
@@ -1,0 +1,2765 @@
+#!/usr/bin/env python3
+# -*- coding: utf-8 -*-
+"""
+Created on Wed Nov 13 15:45:18 2019
+
+@author: alessandro
+"""
+
+#################################################
+### pyHoops: a Python package to web-parse
+### basketball games' play-by-play info
+### and, using basketball data analytics,
+### determine the effect of a player or a lineup
+### on the game
+#################################################
+
+import numpy as np
+import requests
+from bs4 import BeautifulSoup
+import pandas as pd
+import os
+import sys
+import matplotlib.pyplot as plt
+from matplotlib.cbook import get_sample_data
+from matplotlib.offsetbox import OffsetImage, AnnotationBbox
+import random
+import re
+import itertools
+import unicodedata
+from copy import deepcopy
+
+# Input  : string identifying home team, string identifying away team, 
+#          list of available teams
+# Output : index of the logo that represents the home team and away team.
+#          If a team is not recognized in the list of teams available,
+#          it is assigned a default generic logo
+def determine_team_logo(string_homeTeam,string_awayTeam,team_list):
+    idx_logo_homeTeam = np.nan
+    for i in range(0,int(len(team_list))):
+        if team_list[i] in string_homeTeam:
+            idx_logo_homeTeam = i
+            break
+    if np.isnan(idx_logo_homeTeam):
+        idx_logo_homeTeam = int(len(team_list))
+    
+    idx_logo_awayTeam = np.nan
+    for i in range(0,int(len(team_list))):
+        if team_list[i] in string_awayTeam:
+            idx_logo_awayTeam = i
+            break
+    if np.isnan(idx_logo_awayTeam):
+        idx_logo_awayTeam = int(len(team_list))
+        
+    return idx_logo_homeTeam,idx_logo_awayTeam
+
+
+def web_parse_playbyplay(thisUrl): 
+
+    p    = requests.get(thisUrl)
+    
+    if p.status_code < 400:
+        soup   = BeautifulSoup(p.text,'html.parser')
+    
+        home_actions = []        
+        for div in soup.findAll('div',attrs={'class':'action-comment home'}):
+            home_actions.append(div.text.lower())
+            
+        home_score = []        
+        for div in soup.findAll('div',attrs={'class':'scores home'}):
+            home_score.append(div.text)
+            
+        away_actions = []        
+        for div in soup.findAll('div',attrs={'class':'action-comment visitor'}):
+            away_actions.append(div.text.lower())
+            
+        away_score = []        
+        for div in soup.findAll('div',attrs={'class':'scores visitor'}):
+            away_score.append(div.text)
+            
+        time_of_game = []
+        for span in soup.findAll('span',attrs={'class':'time orange'}):
+            time_of_game.append(span.text)
+            
+        quarter_of_game = []
+        for span in soup.findAll('span',attrs={'class':'quarter'}):
+            quarter_of_game.append(span.text.lower())
+            
+        pbp_matrix = []
+        for i in range(0,int(len(home_actions))):
+            this_row = [home_actions[i],home_score[i],time_of_game[i],quarter_of_game[i],
+                        away_score[i],away_actions[i]]
+            pbp_matrix.append(this_row)
+            
+        df_pbp = pd.DataFrame(np.array(pbp_matrix),columns=['Home_action',
+                              'Home_score','time_of_game','quarter_of_game',
[…]
