+++ tests/pyHoops_tests.py
@@ -1,0 +1,281 @@
+#!/usr/bin/env python3
+# -*- coding: utf-8 -*-
+"""
+Created on Wed Nov 13 15:44:28 2019
+
+@author: alessandro
+"""
+
+import os
+import sys
+import pickle
+import numpy as np
+
+sys.path.append('C:/Python27/Lib/site-packages')
+sys.path.append('/Library/TeX/texbin/xelatex')
+sys.path.append('/anaconda/bin/python')
+sys.path.append('/anaconda/lib/python2.7/site-packages')
+sys.path.append('anaconda/envs/py37/lib/python3.7/site-packages')
+
+import pyHoops
+
+#######################
+#######################
+### INPUTS:
+#######################
+#######################
+
+# Number of minutes per quarter. It changes whether European basketball
+# or NBA is considered
+minutes_in_quarter = 10
+
+# Path to current folder
+cwd             = os.getcwd()
+
+# List of teams and associated logos. An additional logo is added, to be associated
+# with teams that do not have a logo in the folder (in this context, logos are
+# specific for the Italian League)
+team_list = ['brescia','brindisi','cantu','cremona','fortitudobologna',
+'milano'	,'pesaro','pistoia','reggioemilia','roma','sassari','trentino',
+'treviso','trieste','varese','venezia','virtusbologna']
+
+team_list_logos = ['Brescia.png','Brindisi.png','Cantu.png','Cremona.png',
+                   'Fortitudo_Bologna.png','Milano.png'	,'Pesaro.png',
+                   'Pistoia.png','Reggio_Emilia.png','Roma.png','Sassari.png',
+                   'Trentino.png','Treviso.png','Trieste.png','Varese.png',
+                   'Venezia.png','Virtus_Bologna.png','LBA_logo.png']
+
+#######################
+#######################
+### TESTS:
+#######################
+#######################
+# Tor un each test, de-comment the four input lines and run the file
+
+## TEST 1
+#thisUrl         = 'http://web.legabasket.it/game/1672570/germani_basket_brescia-openjobmetis_varese_91:74/pbp'
+#thisUrlBoxscore = 'http://web.legabasket.it/game/1672570/germani_basket_brescia-openjobmetis_varese_91:74'
+#string_homeTeam = 'brescia'
+#string_awayTeam = 'varese'
+
+# TEST 2
+thisUrl         = 'http://web.legabasket.it/game/1672568/acqua_s__bernardo_cant__-virtus_roma_74:76/pbp'
+thisUrlBoxscore = 'http://web.legabasket.it/game/1672568/acqua_s__bernardo_cant__-virtus_roma_74:76'
+string_homeTeam = 'cantu'
+string_awayTeam = 'roma'
+
+## TEST 3
+#thisUrl         = 'http://web.legabasket.it/game/1672567/pallacanestro_trieste-happy_casa_brindisi_72:80/pbp'
+#thisUrlBoxscore = 'http://web.legabasket.it/game/1672567/pallacanestro_trieste-happy_casa_brindisi_72:80'
+#string_homeTeam = 'trieste'
+#string_awayTeam = 'brindisi'
+
+## TEST 4
+#thisUrl         = 'http://web.legabasket.it/game/1672569/segafredo_virtus_bologna-d___longhi_treviso_84:79/pbp'
+#thisUrlBoxscore = 'http://web.legabasket.it/game/1672569/segafredo_virtus_bologna-d___longhi_treviso_84:79'
+#string_homeTeam = 'virtusbologna'
+#string_awayTeam = 'treviso'
+
+## TEST 5
[…]
