### 2019-11-14 a231a29
Update pyHoops_tests.py
