### 2019-11-14 #1845 @alessandroBombelli
> Thanks - let us know when this is all done and I'll remove the paused label.

Dear @mado89, @danielskatz, @ctsilva, I updated the Github repo. As mentioned in the Installation section, for now the easiest thing is to download the tests folder as provided, which already contains the pyHoops.py file. I will be soon working on uploading the file in PyPi, so a classic (and more standard) installation of the package is possible. I also updated the readme.md to better show why this package could be relevant for research purposes. The NBA recently launched a competition to use machine learning algorithms to predict game scores using historical data (https://www.nbadatachallenge.com/). This package can be used to extract "real-time" historical data in an easy way (provided the web-scraping part is customized with respect to the actual web-page. Unfortunately there is not free lunch here) for the NBA and other basketball leagues. In the meanwhile, thanks for the patience and do not hesitate to contact me if you have issues.
Cheers,
Alessandro
