### 2019-11-15 alessandroBombelli/pyHoops#56 @mado89
PIP files / structure of the project
Although the installation information in README states the package can be installed using pip the repository does not contain the necessary scripts to make building the package reproducible.

Moreover, the project (and this is related to the point above) is not organised as python projects are usually organised. It seems like the main file pyHoops.py is located in the folder test, the last folder I would search for the file. Please correct me if I am wrong here.

I strongly suggest to reorganise the structure of the project in order to make it easier for other people to use your software. I will try to conduct the review of the software independent from this issue.

As I am no expert in packaging myself I did a quick search and it seems that https://packaging.python.org/tutorials/packaging-projects/ contains all the information necessary and not a lot of work is needed. Completing it would also fix https://github.com/alessandroBombelli/pyHoops/issues/50 and https://github.com/alessandroBombelli/pyHoops/issues/49

https://github.com/openjournals/joss-reviews/issues/1845
