### 2019-11-15 #1845 @mado89
> Dear @mado89, @danielskatz, @ctsilva, I updated the code to reflect the changes in the original website I web-scrape. Tomorrow, I will update the github files so that they be downloaded and run. In the menawhile, I updated the readme.md file as well, and added the MIT license as requested. Thanks for the patience

I am sorry but I can't see the license file. I am looking at the master branch. I guess my problems are also related to the issue I have just created (https://github.com/alessandroBombelli/pyHoops/issues/56) when I inspect the tag "1.5".
In order to not spend too much time digging around. Could I ask you to clean up the repository and notify us once the the issue is fixed? (ie when opening the repository using https://github.com/alessandroBombelli/pyHoops the branch to be reviewed is displayed). Sorry for all the maintenance extra work but I feel your work will benefit from this.
