### 2019-11-15 #1845 @alessandroBombelli
> > Dear @mado89, @danielskatz, @ctsilva, I updated the code to reflect the changes in the original website I web-scrape. Tomorrow, I will update the github files so that they be downloaded and run. In the menawhile, I updated the readme.md file as well, and added the MIT license as requested. Thanks for the patience
>
> I am sorry but I can't see the license file. I am looking at the master branch. I guess my problems are also related to the issue I have just created ([alessandroBombelli/pyHoops#56](https://github.com/alessandroBombelli/pyHoops/issues/56)). When I inspect the tag "1.5" the project looks more like a python package (although LICENSE is located in pyHoops and not in the root directory).
>
> In order to not spend too much time digging around and looking for the code to review, could I ask you to clean up the repository and notify us once this issue is fixed?
> To be precise I'd like to open the repository using https://github.com/alessandroBombelli/pyHoops and see the branch to be reviewed displayed (in case this is not the master branch you can change that somewhere in github afaik). Sorry for all the maintenance extra work but I feel your work will benefit from this.

Hi @mado89, I updated yesterday the repository https://github.com/alessandroBombelli/pyHoops and, more specifically,
(i) I modified the "Installation" section specifying that, for now, the only way to download the file is to download the "tests" folder which contains the file itself
(ii) I added the MIT license in the "License" section (on top of the link to the MIT license website)

As example, my "Installation" section now says

Installation:
The most updated version of pyHoops is directly accessible via the folder tests in this repository. The package pyHoops.py would need to be saved in the same folder as the main code calling it (see Section "Example"). pyHoops will be soon added to PyPi, so that an automatic installation via pip is also possible.

If you are not seeing this, we might be checking different branches. That is my master branch as updated yesterday.
Going back to your advice when it comes to PyPi, you are totally right. I actually followed those very same instructions when I initially created the package, and it worked the first time. I modified the package and wanted to update the version, and all of a sudden the same command I was running from the terminal to upload/update the package in PyPi stopped working. After an unsuccessful afternoon, I gave up. I will follow your advice and try again. If it worked the first time, it should work again. Again, I agree with you in the sense that the current layout is not consistent with what a good package should look like, and I apologize for that
