### 2019-11-15 alessandroBombelli/pyHoops#49 @alessandroBombelli
> Yes, a LICENSE file is needed

Added license file
