### 2019-11-15 #1845 @alessandroBombelli
> This is what I see:
>
> ![Screenshot from 2019-11-15 21-58-11](https://user-images.githubusercontent.com/975497/68975004-63703700-07f2-11ea-8886-946c4ec160e5.png)

That looks right. The "Installation" and "License" sections have been updated though. As example, now I mention that pyHoops will be uploaded to PyPi in the next future and advise the (not so ideal) direct download of the "tests" file. If that is what you see, we are looking at the same thing.

As a side note, there is already a PyPi version of pyHoops (https://pypi.org/project/pyHoops/), which I created following the tutorial you suggested. It is a deprecated version I wanted to update, and that is when I started having issues. I will try to solve the issue during the weekend.

In the meanwhile, thanks for the help and the suggestions.
