### 2019-11-27 #1845 @danielskatz
👋 @mado89 - How is your review proceeding? Specifically, what is blocking you at this point?
