### 2019-12-02 #1845 @mado89
> wave @mado89 - How is your review proceeding? Specifically, what is blocking you at this point?

For me its the general structure of the project. For now it looks more like a utility script that needs to be copied. Although there is a defunct "setup.py" on a different branch. The missing LICENSE file is also on a different branch. So the review would need to consider different branches with different problems.
