### 2019-12-02 #1845 @danielskatz
> For me its the general structure of the project.

What do you suggest for the author to do to improve this?
Can this be reviewed in its current state? (An option is to pause the review until some changes are made.)
