### 2019-12-07 alessandroBombelli/pyHoops#57 @mado89
Logos in charts
I noticed that the charts (the pop up ones, see other issue about the saved ones) have overlaps of team logo and chart.
In some cases the logo even overlaps with the chart title (see Figure_2 for an example).
Can this be fixed?

https://github.com/openjournals/joss-reviews/issues/1845

![Figure_1](https://user-images.githubusercontent.com/975497/70375510-cb500400-18fe-11ea-9130-05302883be67.png)
![Figure_2](https://user-images.githubusercontent.com/975497/70375512-cdb25e00-18fe-11ea-9f44-95f53312b3b8.png)
