### 2019-12-07 alessandroBombelli/pyHoops#58 @mado89
Saved files / output of test empy
Running test/pyHoops_test.py created some image files (e.g. 'fgp_perLineup_brescia.png').
However, they are all empty.
The script did not show me any errors - as far as I can tell.
It this on purpose?

https://github.com/openjournals/joss-reviews/issues/1845
