### 2019-12-07 #1845 @mado89
Review:

I now - finally - had time to give the software a thorough review.
As mentioned in my previous comments I have the feeling that the author needs to polish the structure of the project. Some (vital) aspects such as the LICENSE file and setup.py are spread over different branches.
Moreover, the main software is duplicated (once in root, once in tests).
The generated graphs in the test directory need some work in my opinion. In almost every graph the logo is overlapping with the graph. Sometimes even with the title (I opened an issue for this, https://github.com/alessandroBombelli/pyHoops/issues/57).

Paper:
* The paper describing the software lacks some scientific rigour. The need of the software is only justified by references to popculture or interviews. Especially in sport science what people tell you in interviews differs quite a lot to what is actually done behind the scences (At least in my experience). This is not to say that there is no need for pyHoops.
* The architecture of the software is not described thoroughly. Moreover, I was unable to identify the 'two major blocks' mentioned in the paper when looking at the software.
* The paper mentiones 'indices' but does not describe which indices/metrics are computed and whether and how additional ones could be added to the software.

Software qualitiy:
In my opinion the software could benefit from a refactoring.
As of now it seems that it does not adhere to the principles of object oriented programming.
There are a couple of functions/methods but no checks on input types etc.
Moreover, it seems that webparsing and data processing is tied together which should not be the case.
I'd suggest to have different utility classes for extracting the information from the web and processing it.
This would also solve your problem of changing websites. Moreover, someone could write additional parsers for different websites (if they exist. I haven't checked up on this)

The software seems to be a useful tool. Although, I have to disclose that I do not work in the field of basketball. However, as of now it looks more like a utility script rather than a complete software. I'd suggest improving on the comments made and resubmitting it.

I hope my comments make sense to you @alessandroBombelli If not please let me know so that I can assist in clarifying.
