### 2019-12-07 #1845 @mado89
> @mado89 - Can you make some specific suggestions? (see my previous comment)

I made some comments above and opened some issues. My main concern is that there are some problems with using git / version control in this project (see https://github.com/alessandroBombelli/pyHoops/issues/56 and https://github.com/alessandroBombelli/pyHoops/issues/49).
If I am too strict on this matter please let me know!
