### 2019-12-08 #1845 @danielskatz
Thanks @mado89 - I don't think you are being too strict

One of the requirements for JOSS is

> The software should be feature-complete (no half-baked solutions) and designed for maintainable extension (not one-off modifications). Minor ‘utility’ packages, including ‘thin’ API clients, are not acceptable

I think the things you are asking for here fall into the
> designed for maintainable extension (not one-off modifications)

category
