### 2019-12-09 #1845 @alessandroBombelli
> Review:
>
> I now - finally - had time to give the software a thorough review.
> As mentioned in my previous comments I have the feeling that the author needs to polish the structure of the project. Some (vital) aspects such as the LICENSE file and setup.py are spread over different branches.
> Moreover, the main software is duplicated (once in root, once in tests).
> The generated graphs in the test directory need some work in my opinion. In almost every graph the logo is overlapping with the graph. Sometimes even with the title (I opened an issue for this, [alessandroBombelli/pyHoops#57](https://github.com/alessandroBombelli/pyHoops/issues/57)).
>
> Paper:
>
> * The paper describing the software lacks some scientific rigour. The need of the software is only justified by references to popculture or interviews. Especially in sport science what people tell you in interviews differs quite a lot to what is actually done behind the scences (At least in my experience). This is not to say that there is no need for pyHoops.
> * The architecture of the software is not described thoroughly. Moreover, I was unable to identify the 'two major blocks' mentioned in the paper when looking at the software.
> * The paper mentiones 'indices' but does not describe which indices/metrics are computed and whether and how additional ones could be added to the software.
>
> Software qualitiy:
> In my opinion the software could benefit from a refactoring.
> As of now it seems that it does not adhere to the principles of object oriented programming.
> There are a couple of functions/methods but no checks on input types etc.
> Moreover, it seems that webparsing and data processing is tied together which should not be the case.
> I'd suggest to have different utility classes for extracting the information from the web and processing it.
> This would also solve your problem of changing websites. Moreover, someone could write additional parsers for different websites (if they exist. I haven't checked up on this)
>
> The software seems to be a useful tool. Although, I have to disclose that I do not work in the field of basketball. However, as of now it looks more like a utility script rather than a complete software. I'd suggest improving on the comments made and resubmitting it.
>
> I hope my comments make sense to you @alessandroBombelli If not please let me know so that I can assist in clarifying.

Dear @mado89, thanks for the exhaustive comments. Indeed, I agree with all of them and already started to apply changes. Here are my replies to your comments:

- I worked on the PyPi implementation (so that this becomes an actual package) over the weekend, but had some issues in the uploading process. I think I am following the right steps (I am following these intructions https://medium.com/@joel.barmettler/how-to-upload-your-python-package-to-pypi-65edc5fe9c56), but there is still something wrong. This is indeed at the top of my priority queue

- Related to the above issue, I will fix the structure of the GitHub page to have more consistency (eliminating repeated files etc), and provide a better description of the architecture

- Webparsing vs data analysis: I already added a first update. Now, users can decide to webparse a webpage, or to use a .pickle or .xlsx file with the processed play-by-play structure. These are the two blocks I am referring to: (i) webparsing (if necessary), and (ii) data analysis. As you pointed out, I should be more clear in differentiating them. An user might want to perform new/different analyses on an already present dataset (hence skipping block (i)), or perform new analyses from scratch (hence using (i) and (ii) in sequence). A better division of the two blocks would also benefit users interested in different leagues, where play-by-play have a different structures and jargon (block (i) should be changed accordingly). If they are stored as pyHoops requires (I should provide some better instructions in this sense) block (ii) is applicable to any dataset

- I am aware in some Figures the logo might be too big or partially overlap with the title. Logos differ in size quite considerably, hence either the user must specify a different initial coordinate set for the logo, or I need to figure out something that can detect overlapping cases and automatically adjust the position/size. Indeed, this latter option is advisable

- A last comment on the applicability/scientific relevance. I agree there is not much literature on the topic (yet) if compared to other sports. Given this "big data" era, an increasing interest towards basketball has started recently. See this competition that was just launched by the NBA  ---> https://nbadatachallenge.com/

All in all, I want to join @danielskatz in thanking you for your thorough comments. I wil ldo my best to implement all of them. If something in my replies is vague or not clear, please let me know.

Cheers,

Alessandro
