### 2020-01-06 #1845 @danielskatz
👋 @alessandroBombelli  - Can you give us a post-holiday status update?
