### 2020-02-24 #1845 @alessandroBombelli
Dear Dr. Katz,
I apologize, but I was lagging behind with my research work at TU Delft,
and I had to further delay this work. Being a passion/hobby, it was the
first thing to cut. It is taking too long, please either reject the paper
or prolong the "pause" option. Given the time-frame, I will totally
understand a rejection (on top of the other shortcoming of the work in its
current form).

Again, apologies. I will improve the work as soon as I can and either
submit an improved version or re-submit it depending on the decision
Best,

Alessandro

On Mon, Jan 6, 2020 at 3:23 PM Daniel S. Katz <notifications@github.com>
wrote:

> We will wait until then - I will mark this as paused in the meantime.
>
